requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
PyJWT==2.8.0
redis==5.0.4
langchain==0.2.0
python-dotenv==1.0.1
langchain-openai==0.1.7
pymongo==4.7.2
langchain-community==0.2.0
langchainhub==0.1.15
gunicorn==22.0.0
requests==2.32.3
cockroachdb==0.3.5
SQLAlchemy==1.4.5
psycopg2-binary==2.9.9
flask-sock==0.7.0
geopy==2.4.1
pika==1.3.2
docx2txt==0.8
assemblyai==0.30.0
numpy==1.24.4
neo4j==5.23.1
sympy==1.13.1
boto3==1.35.8
botocore==1.35.9
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
flake8==7.1.1
black==24.8.0
flake8-bugbear==24.8.19
flake8-import-order==0.18.2
tiktoken==0.7.0
openai==1.47.1
opencv-python==4.10.0.84
pillow==10.4.0
pytesseract==0.3.13
transformers==4.41.0
torch==2.2.2
werkzeug==3.0.4
pinecone==5.3.1
